requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
requests>=2.31.0
hypothesis>=6.92.0
